% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\docType{data}
\name{demo}
\alias{demo}
\title{A simulated spatial transcriptomics dataset}
\format{
a Seurat object with 22,127 genes measured in 100 spatial barcodes
}
\usage{
data(demo)
}
\description{
A toy simulated spatial transcriptomics dataset, used to demonstrate the
functionality and usage of Magellan.
}
\keyword{datasets}
