% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigate_space.R
\name{navigate_space}
\alias{navigate_space}
\title{Perform spatial prioritization of the response to a biological perturbation}
\usage{
navigate_space(
  input,
  meta = NULL,
  coords = NULL,
  k = 50,
  label_col = "label",
  coord_cols = c("coord_x", "coord_y"),
  n_subsamples = 50,
  subsample_size = 20,
  folds = 3,
  var_quantile = 0.5,
  feature_perc = 0.5,
  n_threads = 32,
  show_progress = T,
  augur_mode = c("default", "velocity"),
  classifier = c("rf", "lr"),
  rf_params = list(trees = 100, mtry = 2, min_n = NULL, importance = "accuracy"),
  lr_params = list(mixture = 1, penalty = "auto")
)
}
\arguments{
\item{input}{a matrix, data frame, or \code{Seurat} object containing
gene expression values (genes in rows, cells in columns) and,
optionally, metadata about each spatial barcode}

\item{meta}{optionally, a data frame containing metadata about the
\code{input} gene-by-barcode matrix, at minimum containing the label
associated with each barcode
(e.g., group, disease, timepoint); can be left as
\code{NULL} if \code{input} is a \code{Seurat} object}

\item{coords}{optionally, a data frame containing the spatial coordinates
for each barcode in the \code{input} gene-by-barcode matrix; can be left as
\code{NULL} if \code{input} is a \code{Seurat} object}

\item{k}{the number of spatial nearest-neighbors to use in the AUC
calculation; defaults to \code{50}}

\item{label_col}{the column of the \code{meta} data frame, or the
metadata container in the \code{Seurat} object, that
contains condition labels (e.g., disease, timepoint) for each barcode in the
gene-by-barcode expression matrix; defaults to \code{"label"}}

\item{coord_cols}{the names of the columns in the \code{coords} data frame,
or the metadata container in the \code{Seurat} object, that contain the
coordinates of each spatial barcode in the gene-by-barcode expression
matrix; defaults to \code{c("coord_x", "coord_y")}}

\item{n_subsamples}{the number of times to repeat the cross-validation
procedure for each barcode; defaults to \code{50}.
Set to \code{0} to omit subsampling altogether,
calculating performance on the entire dataset, but note that this may
introduce bias due to cell type or label class imbalance.
Note that when setting \code{augur_mode = "permute"}, values less than
\code{100} will be replaced with a default of \code{500}.}

\item{subsample_size}{the number of barcodes to randomly sample from among
the nearest neighbors in each iteration of the cross-validation procedure;
cannot be greater than \code{k}}

\item{folds}{the number of folds of cross-validation to run; defaults to
\code{3}}

\item{var_quantile}{the quantile of highly variable genes to retain using
the variable gene filter (\link[Augur]{select_variance});
defaults to \code{0.5}}

\item{feature_perc}{the proportion of genes that are randomly selected as
features for input to the classifier in each subsample using the
random gene filter (\link[Augur]{select_random}); defaults to \code{0.5}}

\item{n_threads}{the number of threads to use for parallelization;
defaults to \code{32}.}

\item{show_progress}{if \code{TRUE}, display a progress bar for the analysis
with estimated time remaining}

\item{augur_mode}{one of \code{"default"} or \code{"velocity"}.
Setting \code{augur_mode = "velocity"} disables feature
selection, assuming feature selection has been performed by the RNA
velocity procedure to produce the input matrix}

\item{classifier}{the classifier to use in calculating area under the curve,
one of \code{"rf"} (random forest) or \code{"lr"} (logistic regression);
defaults to \code{"rf"}, which is the recommended setting}

\item{rf_params}{for \code{classifier} == \code{"rf"}, a list of parameters
for the random forest models, containing the following items (see
\link[parsnip]{rand_forest} from the \code{parsnip} package):
\describe{
\item{"mtry"}{the number of features randomly sampled at each split
in the random forest classifier; defaults to \code{2}}
\item{"trees"}{the number of trees in the random forest classifier;
defaults to \code{100}}
\item{"min_n"}{the minimum number of observations to split a node in the
random forest classifier; defaults to \code{NULL}}
\item{"importance"}{the method of calculating feature importances
to use; defaults to \code{"accuracy"}; can also specify \code{"gini"}}
}}

\item{lr_params}{for \code{classifier} == \code{"lr"}, a list of parameters
for the logistic regression models, containing the following items (see
\link[parsnip]{logistic_reg} from the \code{parsnip} package):
\describe{
\item{"mixture"}{the proportion of L1 regularization in the model;
defaults to \code{1}}
\item{"penalty"}{the total amount of regularization in the model;
defaults to \code{"auto"}, which uses \link[glmnet]{cv.glmnet} to set
the penalty}
}}
}
\value{
a list of class \code{"Magellan"}, containing the following items:
\enumerate{
\item \code{parameters}: the parameters provided to this function as input
\item \code{results}: the area under the curve for each barcode, in each
fold, in each subsample, in the comparison of interest, as well as a
series of other classification metrics
\item \code{AUC}: a summary of the mean AUC for each barcode (for
continuous experimental conditions, this is replaced by a \code{CCC}
item that records the mean concordance correlation coefficient for each
barcode)
}
}
\description{
Prioritize spatial locations involved in a complex biological process by
training a machine-learning model to predict sample labels
(e.g., disease vs. control, treated vs. untreated, or time post-stimulus),
and evaluate the performance of the model in cross-validation.
}
\details{
If a \code{Seurat} object is provided as input, Magellan will use the default
assay (i.e., whatever \link[Seurat]{GetAssayData} returns) as input. To
use a different assay, provide the expression matrix and metadata as input
separately, using the \code{input} and \code{meta} arguments.
Additionally, Magellan will assume the coordinates of the spatial barcodes
can be found in \code{input@images$slice1@coordinates}. To override this,
specify the count matrix, metadata, and coordinates separately.
}
